fix type of variable

+++ cfg.cpp
@@ -47,7 +47,7 @@ const ini_var_t ini_vars[] = {
 	{ "RBF_HIDE_DATECODE", (void*)(&(cfg.rbf_hide_datecode)), UINT8, 0, 1, 1 },
 	{ "MENU_PAL", (void*)(&(cfg.menu_pal)), UINT8, 0, 1, 1 },
 	{ "BOOTCORE", (void*)(&(cfg.bootcore)), STRING, 0, sizeof(cfg.bootcore) - 1, 1 },
-	{ "BOOTCORE_TIMEOUT", (void*)(&(cfg.bootcore_timeout)), INT8, 10, 30, 1 },
+	{ "BOOTCORE_TIMEOUT", (void*)(&(cfg.bootcore_timeout)), INT16, 10, 30, 1 },
 };
 
 // mist ini config
